xv6 files
xv6 is a great Unix like OS for learning OS concepts. Xv6 is a re-implementation of the Unix sixth edition in order to use as a learning tool. xv6 was developed by MIT as a teaching operating system for their “6.828” course.
 You can get a copy of full source code from this Git Repo: https://github.com/mit-pdos/xv6-public .

You can compile it and run it using an emulator like qemu.